Extract functions in new file
